changed figure 8 and 9
