+++ lineage/figures/figure8.py
@@ -1,49 +1,54 @@
 """
-File: figure7.py
-Purpose: Generates figure 7.
+File: figureS11.py
+Purpose: Generates figure S11.
 
 AIC.
 """
 import numpy as np
+from matplotlib.ticker import MaxNLocator
+
+from ..Analyze import run_Analyze_AIC
+from ..LineageTree import LineageTree
 
 # States to evaluate with the model
 from ..states.StateDistributionGaPhs import StateDistribution
 
-from .figureCommon import getSetup, subplotLabel
-from .figureS11 import run_AIC, figure_maker
+from .figureCommon import getSetup, lineage_good_to_analyze, subplotLabel
 
+# TODO
+# Calculate A-AIC where you removed the count of elements in T and pi that are zero, and add for those that are 1.
 
 desired_num_states = np.arange(1, 8)
+
+# Setting up state distributions and E
+Sone = StateDistribution(0.99, 0.9, 10, 2, 10, 2)
+Stwo = StateDistribution(0.9, 0.9, 20, 3, 20, 3)
+Sthree = StateDistribution(0.85, 0.9, 30, 4, 30, 4)
+Sfour = StateDistribution(0.8, 0.9, 40, 5, 40, 5)
+Eone = [Sone, Sone]
+Etwo = [Sone, Stwo]
+Ethree = [Sone, Stwo, Sthree]
+Efour = [Sone, Stwo, Sthree, Sfour]
+E = [Eone, Etwo, Ethree, Efour, Eone, Etwo, Ethree, Efour]
 
 
 def makeFigure():
     """
-    Makes figure 7.
+    Makes figure 10.
     """
     ax, f = getSetup((12, 6), (2, 4))
-
-    # Setting up state distributions and E
-    Sone = StateDistribution(0.99, 0.9, 10, 2, 10, 2)
-    Stwo = StateDistribution(0.9, 0.9, 20, 3, 20, 3)
-    Sthree = StateDistribution(0.85, 0.9, 30, 4, 30, 4)
-    Sfour = StateDistribution(0.8, 0.9, 40, 5, 40, 5)
-    Eone = [Sone, Sone]
-    Etwo = [Sone, Stwo]
-    Ethree = [Sone, Stwo, Sthree]
-    Efour = [Sone, Stwo, Sthree, Sfour]
-    E = [Eone, Etwo, Ethree, Efour, Eone, Etwo, Ethree, Efour]
-
     # making lineages and finding AICs (assign number of lineages here)
     AIC = [run_AIC(.1, e, 10, idx > 3) for idx, e in enumerate(E)]
 
     # Finding proper ylim range for all 4 uncensored graphs and rounding up
-    upper_ylim_uncensored = int(1 + max(np.max(np.ptp(AIC[0], axis=0)), np.max(np.ptp(
-        AIC[1], axis=0)), np.max(np.ptp(AIC[2], axis=0)), np.max(np.ptp(AIC[3], axis=0))) / 25.0) * 25
-
-    # Finding proper ylim range for all 4 censored graphs and rounding up
-    upper_ylim_censored = int(1 + max(np.max(np.ptp(AIC[4], axis=0)), np.max(np.ptp(
-        AIC[5], axis=0)), np.max(np.ptp(AIC[6], axis=0)), np.max(np.ptp(AIC[7], axis=0))) / 25.0) * 25
-
+    upper_ylim_uncensored = int(1 + max(np.max(np.ptp(AIC[0], axis=0)),
+                              np.max(np.ptp(AIC[1], axis=0)),
+                              np.max(np.ptp(AIC[2], axis=0)),
+                              np.max(np.ptp(AIC[3], axis=0))) / 25.0) * 25
+    upper_ylim_censored = int(1 + max(np.max(np.ptp(AIC[4], axis=0)),
+                              np.max(np.ptp(AIC[5], axis=0)),
+                              np.max(np.ptp(AIC[6], axis=0)),
+                              np.max(np.ptp(AIC[7], axis=0))) / 25.0) * 25
     upper_ylim = [upper_ylim_uncensored, upper_ylim_censored]
 
     # Plotting AICs
@@ -51,5 +56,66 @@ def makeFigure():
         figure_maker(ax[idx], a, (idx % 4) + 1,
                      upper_ylim[int(idx / 4)], idx > 3)
     subplotLabel(ax)
-
     return f
+
+
+def run_AIC(relative_state_change, E, num_lineages_to_evaluate=10, censored=False):
+    """
+    Run's AIC for known lineages with known pi,
+    and T values and stores the output for
+    figure drawing.
+    """
+    # Setting up pi and Transition matrix T:
+    #   pi: All states have equal initial probabilities
+    #   T:  States have high likelihood of NOT changing, with frequency of change determined mostly by the relative_state_change variable
+    #   (as relative state change -> inf state change probabilities approach equality)
+    pi = np.ones(len(E)) / len(E)
+    T = (np.eye(len(E)) + relative_state_change)
+    T = T / np.sum(T, axis=1)[:, np.newaxis]
+
+    # Creating censored lineages
+    if censored:
+        lineages = [LineageTree.init_from_parameters(
+            pi, T, E, 2**6 - 1, censor_condition=3, experiment_time=1200) for _ in range(num_lineages_to_evaluate)]
+
+    # Creating uncensored lineages
+    else:
+        lineages = [LineageTree.init_from_parameters(
+            pi, T, E, 2**6 - 1) for _ in range(num_lineages_to_evaluate)]
+    lineages = [l for l in lineages if lineage_good_to_analyze(l)]
+
+    # Storing AICs into array
+    AICs = np.empty(len(desired_num_states))
+    output = run_Analyze_AIC(lineages, desired_num_states)
+    for idx in range(len(desired_num_states)):
+        AICs[idx], _ = output[idx][0].get_AIC(output[idx][2], None)
+
+    return AICs
+
+
+def figure_maker(ax, AIC_holder, true_state_no, upper_ylim, censored=False):
+    """
+    Makes figure 11.
+    """
+    # Normalizing AIC
+    AIC_holder = AIC_holder - np.min(AIC_holder)
+
+    # Creating Histogram and setting ylim
+    ax2 = ax.twinx()
+    ax2.set_ylabel("Lineages Predicted")
+    ax2.hist(np.argmin(AIC_holder) + 1, rwidth=1,
+             alpha=.2, bins=desired_num_states, align='left')
+    ax2.set_yticks(np.linspace(0, len(AIC_holder), 5))
+
+    # Creating AIC plot and matching gridlines
+    ax.set_xlabel("Number of States Predicted")
+    ax.plot(desired_num_states, AIC_holder, "k", alpha=0.5, linewidth=2)
+    ax.set_ylabel("Normalized AIC")
+    ax.set_yticks(np.linspace(0, upper_ylim, len(ax2.get_yticks())))
+    ax.xaxis.set_major_locator(MaxNLocator(integer=True))
+
+    # Adding title
+    title = "Censored " if censored else ""
+    title += f"AIC Under {true_state_no} True "
+    title += "States" if true_state_no != 1 else "State"
+    ax.set_title(title)
